Add an animated ghost on the screen :)

+++ project-egg-core/src/Physics/Transform.cpp
@@ -1,12 +1,12 @@
 #include "Transform.h"
 
-Transform::Transform(float x, float y):X(x),Y(y)
+Transform::Transform(int x, int y):X(x),Y(y)
 {
 }
 
 void Transform::Log(const char* message)
 {
-	SDL_Log("%s (X, Y) = (%f, %f)\n", message, X, Y);
+	SDL_Log("%s (X, Y) = (%d, %d)\n", message, X, Y);
 }
 
 void Transform::Translate(Vector2D* v)
